feat: show branches sorted by date
This was quite a journey: The rest api and graphql calls were so slow. And, they provide the branches only in alphabetical
order. Therefore I looked around and found a very fast approach by cloning the repository and only download the stuff
I need. This is very fast and normally after a couple of seconds you have the complete result, ordered by last activity.